Extract tabular comparison views to their own module.
