+++ webapp/chlamdb/urls.py
@@ -2,7 +2,8 @@
 from django.urls import re_path
 from django.views.generic import TemplateView
 from django.views.generic.base import RedirectView
-from views import entry_lists, fam, hits_extraction, venn, views
+from views import (entry_lists, fam, hits_extraction, tabular_comparison, venn,
+                   views)
 
 favicon_view = RedirectView.as_view(url='/assets/favicon.ico', permanent=True)
 
@@ -19,11 +20,11 @@ urlpatterns = [
     re_path(r'^plot_region/$', views.plot_region, name="plot_region"),
     re_path(r'^plot_heatmap/([a-zA-Z0-9_\-]+)', views.plot_heatmap, name="plot_heatmap"),  # noqa
     re_path(r'^phylogeny', views.phylogeny, name='phylogeny'),
-    re_path(r'^pfam_comparison', views.PfamComparisonView.as_view(), name="pfam_comparison"),  # noqa
+    re_path(r'^pfam_comparison', tabular_comparison.PfamComparisonView.as_view(), name="pfam_comparison"),  # noqa
     re_path(r'^pan_genome/([a-zA-Z0-9_]+)', views.pan_genome, name="pan_genome"),  # noqa
     re_path(r'^orthogroup_list_cog_barchart/([a-zA-Z0-9_]+)/$', views.orthogroup_list_cog_barchart, name="orthogroup_list_cog_barchart"),  # noqa
     re_path(r'^orthogroup_list_cog_barchart/$', views.orthogroup_list_cog_barchart, name="orthogroup_list_cog_barchart"),  # noqa
-    re_path(r'^orthogroup_comparison', views.OrthogroupComparisonView.as_view(), name="orthogroup_comparison"),  # noqa
+    re_path(r'^orthogroup_comparison', tabular_comparison.OrthogroupComparisonView.as_view(), name="orthogroup_comparison"),  # noqa
     re_path(r'^orthogroup/([a-zA-Z0-9_\.\-]+)', views.orthogroup, name="orthogroup"),  # noqa
     re_path(r'^module_comparison/$', views.module_comparison, name="module_comparison"),  # noqa
     re_path(r'^module_cat_info/([a-zA-Z0-9_\.]+)/([a-zA-Z0-9_\.\+-]+)$', views.module_cat_info, name="module_cat_info"),  # noqa
@@ -31,7 +32,7 @@ urlpatterns = [
     re_path(r'^locusx/([a-zA-Z0-9_\.\-]+)', views.locusx, name="locusx"),
     re_path(r'^locusx$', views.locusx, name="locusx"),
     re_path(r'^ko_venn_subset/([a-zA-Z0-9_\.\+-]+)$', venn.VennKoSubsetView.as_view(), name="ko_venn_subset"),  # noqa
-    re_path(r'^ko_comparison', views.KoComparisonView.as_view(), name="ko_comparison"),  # noqa
+    re_path(r'^ko_comparison', tabular_comparison.KoComparisonView.as_view(), name="ko_comparison"),  # noqa
     re_path(r'^ko_barchart/$', views.ko_barchart, name="ko_barchart"),
     re_path(r'^kegg_module_subcat$', views.kegg_module_subcat, name="kegg_module_subcat"),  # noqa
     re_path(r'^KEGG_module_map/([a-zA-Z0-9_\.]+)$', views.KEGG_module_map, name="KEGG_module_map"),  # noqa
@@ -66,12 +67,12 @@ urlpatterns = [
     re_path(r'^entry_list_amr$', entry_lists.AmrEntryListView.as_view(), name="entry_list_amr"),  # noqa
     re_path(r'^cog_venn_subset/([A-Z])$', venn.VennCogSubsetView.as_view(), name="cog_venn_subset"),  # noqa
     re_path(r'^cog_phylo_heatmap/([a-zA-Z0-9_\-]+)', views.cog_phylo_heatmap, name="cog_phylo_heatmap"),  # noqa
-    re_path(r'^cog_comparison', views.CogComparisonView.as_view(), name="cog_comparison"),  # noqa
+    re_path(r'^cog_comparison', tabular_comparison.CogComparisonView.as_view(), name="cog_comparison"),  # noqa
     re_path(r'^cog_barchart/$', views.cog_barchart, name="cog_barchart"),
     re_path(r'^circos_main/$', views.circos_main, name="circos_main"),
     re_path(r'^circos/$', views.circos, name="circos"),
     re_path(r'^blast/$', views.blast, name="blast"),
-    re_path(r'^amr_comparison', views.AmrComparisonView.as_view(), name="amr_comparison"),  # noqa
+    re_path(r'^amr_comparison', tabular_comparison.AmrComparisonView.as_view(), name="amr_comparison"),  # noqa
     re_path(r'^about$', views.about, name="about"),
     re_path(r'^.*$', views.home, name="home"),
     re_path('^robots.txt$', TemplateView.as_view(template_name='robots.txt', content_type='text/plain')),  # noqa
